+++ RELEASE_NOTES.md
@@ -4,7 +4,8 @@
 
 This release is in development.
 
-- Added functions current-dateTime(), dateTime(), format-dateTime(), millis-to-dateTime(), timestamp().
+- Added functions current-dateTime(), dateTime(), dateTime-to-millis(), format-dateTime(), 
+  millis-to-dateTime(), parse-dateTime(), timestamp().
 
 ## Compatibility